feat: almost ready
