+++ examples/conversational-ai/talk-to-santa/app/layout.tsx
@@ -15,8 +15,8 @@ const geistMono = localFont({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Talk to Santa | ElevenLabs",
+  description: "Have a magical conversation with Santa Claus, powered by Conversational AI.",
 };
 
 export default function RootLayout({
