+++ examples/conversational-ai/talk-to-santa/package.json
@@ -18,6 +18,7 @@
     "@react-three/drei": "^9.103.0",
     "@react-three/fiber": "9.0.0-alpha.8",
     "@vercel/analytics": "^1.4.1",
+    "@vercel/blob": "^0.26.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "elevenlabs": "^0.18.1",
@@ -25,13 +26,15 @@
     "lottie-react": "^2.4.0",
     "lucide-react": "^0.462.0",
     "next": "15.0.3",
+    "next-safe-action": "^7.9.9",
     "react": "19.0.0-rc-66855b96-20241106",
     "react-dom": "19.0.0-rc-66855b96-20241106",
     "react-snowfall": "^2.2.0",
     "tailwind-merge": "^2.5.5",
     "tailwindcss-animate": "^1.0.7",
     "three": "^0.160.0",
-    "vaul": "^1.1.1"
+    "vaul": "^1.1.1",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@types/node": "^20",
